rustchain-telegram-bot/python/requirements.txt
python-telegram-bot==21.0
web3==6.15.1
python-dotenv==1.0.1
aiohttp==3.9.3
